requirements.txt
black==23.7.0
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
SQLAlchemy==2.0.20
requests==2.31.0
twilio==8.7.0
django-extensions==3.2.3
django-redis==5.3.0
pretty-errors==1.2.25
openpyxl==3.0.10